+++ CHANGELOG.md
@@ -1,5 +1,62 @@
 # Changelog
 
+## 2023.01.03
+
+
+### New
+
+- Add two function to **Set** and **Remove** Restic enivronment
+  - Yes, environment variables are used now instead of files or script variable. You can import this module in a PowerShell session to use it manually!
+- Script is loclized now! Your language is french? Script is in french now!
+  - Precisely, the script will manage en-US and fr-FR. Feel free to ask me for a new language!
+
+### Change
+
+- Rename `func` folder into `modules`
+- Replace Microsoft cmdlet `Start-Process` by the custom command `Start-Command` which use .Net
+- Using a *json* file instead of *ini* file
+  - If the file *settings.json* does not exist, the script will ask questions to create it! Thinking for people don't know (or like 😋) the json!
+  - The setting file is test before run to be sure everything is consistent
+
+### Fix
+
+- Some typo
+
+### Log
+
+- *090c60a* - chore(img): demo gif
+- *a83ddc0* - feat(Get-ResticGameSnapshots): use RootPath parameter for New-Settings
+- *59a2e48* - feat(Clean-Restic): use RootPath parameter for New-Settings
+- *cd5fc42* - fix(defaultsettings var): fix messages variable name
+- *9806965* - fix(new-settings): add RootPath parameter and fix messages variable name
+- *4d4a3e2* - fix(set-environment): set variable name
+- *001f020* - fix(local): fix variable name and add variable for New-Settings.psd1
+- *7dfc195* - feat(local): replace all string by localized string
+- *0c4c6d8* - feat(local): missing string added
+- *f0b19b9* - fix(stats): fix a typo issue in the name of the file for the script ConvertTo-HashtableSize.ps1
+- *715cbbf* - feat(local): all string in settings functions has been localized!
+- *9e885eb* - feat(local): import localized data. Need to replace each string in localized data file by the variable
+- *3319c08* - feat(local): all localized string for modules and Clean-Restic Only en-US and fr-FR for the moment !
+- *2d62cdc* - refactor(local): Use script name for the message
+- *8460b72* - feat(Clean & Get): Implement 'Test-Settings' Implicitly by using Get-Settings
+- *555f1f7* - feat(environment): move Set/New-Environment to be used outside this script
+- *93ed3b2* - feat(settings module): Add 'Test-Settings'
+- *ee91f7b* - chore(git): ignore personal folder
+- *8503584* - feat(message): Prepare for localized message File for en-US Culture ready, fr-FR to do.
+- *153bb10* - feat(message): Prepare for localized message Replace all '>>' by 'Out-Null $logFile -Encoding utf8 -Append'
+- *30a74b8* - fix(Clean-Restic): Change a message for a future Search'n'Replace
+- *0e6ffa6* - chore(git): add a personal working folder to .gitignore
+- *16a4dde* - refactor(Get-ResticGameSnapshots): remove warning on variable set but not used
+- *af5ff79* - feat(settings module): function to create settings
+- *79d117e* - feat(settings): use a the new settings variable
+- *8180ff7* - feat(settings file): use a json file instead of an ini file
+- *788e816* - doc: file that contain an example of the prune command
+- *384b06b* - feat(Clean-Restic): replace all 'Start-Process' by 'Start-Command' (custom)
+- *38abd88* - fix(env): literal path to settings file
+- *99663e0* - feat(env): using env var now! Easier to use Restic like this!
+- *be6e4f2* - feat(modules): add two functions to set and remove Restic env var
+- *931a2e7* - chore(modules): move modules from 'func' to 'modules' subfolders
+
 ## 2022.12.13
 
 ### Change